Added Templates and Saving Animations
You can now save animations using the watch_sand method. This method is no longer static and now works by repeatedly adding a passed template sandpile to an empty one of a given size. This will happen over a user-defined number of iterations, after which the animation will be shown or saved, depending on the options chosen.
Templates provide auto-generation of a few simple sandpiles to try out with the watch_sand animation.